Add files via upload
The goal of this application is to log your workout. Select the location on the map where you want to work out then on the form where you can input your workout, in the sidebar, fill the details then the info gets logged in the location. You can also log multiple workouts in different locations. The application allows you to log cycling (Elev Gain, meters) and running workouts (Cadence, steps/min).When you click on the workout in the sidebar, it will move you to the workout lcation in the map.

When you close the map and reload it, the application will keep its state

The map is loaded from a third party service and the position is automatically obtained using the geolocation.